[Chore] 이름 및 경로 변경
1. RedisUser 삭제
2. RedisUser EmbbedableType, EnumType 경로 변경
