+++ src/main/java/com/tools/seoultech/timoproject/matching/domain/board/entity/embeddableType/CompactPlayerHistory.java
@@ -1,4 +1,4 @@
-package com.tools.seoultech.timoproject.matching.domain.user.entity.embeddableType;
+package com.tools.seoultech.timoproject.matching.domain.board.entity.embeddableType;
 
 import com.tools.seoultech.timoproject.riot.dto.MatchSummaryDTO;
 import com.tools.seoultech.timoproject.riot.dto.RankInfoDto;
