+++ src/main/java/com/tools/seoultech/timoproject/matching/domain/board/entity/embeddableType/UserInfo.java
@@ -1,0 +1,22 @@
+package com.tools.seoultech.timoproject.matching.domain.board.entity.embeddableType;
+
+import com.tools.seoultech.timoproject.matching.domain.board.entity.enumType.PlayCondition;
+import com.tools.seoultech.timoproject.matching.domain.board.entity.enumType.PlayPosition;
+import com.tools.seoultech.timoproject.matching.domain.board.entity.enumType.PlayStyle;
+import com.tools.seoultech.timoproject.matching.domain.board.entity.enumType.VoiceChat;
+import jakarta.persistence.Embeddable;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Embeddable
+@Getter
+@AllArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class UserInfo {
+    private PlayPosition myPosition;
+    private PlayStyle myStyle;
+    private PlayCondition myStatus;
+    private VoiceChat myVoice;
+}
