+++ src/main/java/com/tools/seoultech/timoproject/matching/domain/board/entity/enumType/Gender.java
@@ -1,0 +1,7 @@
+package com.tools.seoultech.timoproject.matching.domain.board.entity.enumType;
+
+public enum Gender {
+    MALE,
+    FEMALE,
+    SECRET
+}
