+++ src/main/java/com/tools/seoultech/timoproject/matching/domain/board/entity/enumType/PlayCondition.java
@@ -1,0 +1,7 @@
+package com.tools.seoultech.timoproject.matching.domain.board.entity.enumType;
+
+public enum PlayCondition {
+    FIRST,
+    CONTINUE,
+    LAST
+}
