+++ src/main/java/com/tools/seoultech/timoproject/matching/domain/board/entity/enumType/PlayPosition.java
@@ -1,4 +1,4 @@
-package com.tools.seoultech.timoproject.matching.domain.user.entity.enumType;
+package com.tools.seoultech.timoproject.matching.domain.board.entity.enumType;
 
 public enum PlayPosition {
     NOTHING,
