+++ src/main/java/com/tools/seoultech/timoproject/matching/domain/board/entity/enumType/PlayStyle.java
@@ -1,0 +1,7 @@
+package com.tools.seoultech.timoproject.matching.domain.board.entity.enumType;
+
+public enum PlayStyle {
+    NONE,
+    FUN,
+    HARDCORE,
+}
