+++ src/main/java/com/tools/seoultech/timoproject/matching/domain/board/entity/enumType/VoiceChat.java
@@ -1,0 +1,6 @@
+package com.tools.seoultech.timoproject.matching.domain.board.entity.enumType;
+
+public enum VoiceChat {
+    ENABLED,
+    DISABLED,
+}
