+++ src/main/java/com/tools/seoultech/timoproject/matching/domain/myPage/dto/MyPageDTO.java
@@ -1,12 +1,9 @@
 package com.tools.seoultech.timoproject.matching.domain.myPage.dto;
 
 import com.tools.seoultech.timoproject.matching.domain.board.dto.BoardDTO;
-import com.tools.seoultech.timoproject.matching.domain.board.entity.redis.RedisBoard;
 import com.tools.seoultech.timoproject.matching.domain.myPage.entity.EnumType.MatchingCategory;
 import com.tools.seoultech.timoproject.matching.domain.myPage.entity.EnumType.MatchingStatus;
-import com.tools.seoultech.timoproject.matching.domain.myPage.entity.redis.RedisMyPage;
 import com.tools.seoultech.timoproject.matching.domain.user.dto.UserDTO;
-import com.tools.seoultech.timoproject.matching.domain.user.entity.redis.RedisUser;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Builder;
