--- src/main/java/com/tools/seoultech/timoproject/matching/domain/user/entity/embeddableType/UserInfo.java
@@ -1,22 +1,0 @@
-package com.tools.seoultech.timoproject.matching.domain.user.entity.embeddableType;
-
-import com.tools.seoultech.timoproject.matching.domain.user.entity.enumType.PlayCondition;
-import com.tools.seoultech.timoproject.matching.domain.user.entity.enumType.PlayPosition;
-import com.tools.seoultech.timoproject.matching.domain.user.entity.enumType.PlayStyle;
-import com.tools.seoultech.timoproject.matching.domain.user.entity.enumType.VoiceChat;
-import jakarta.persistence.Embeddable;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Embeddable
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class UserInfo {
-    private PlayPosition myPosition;
-    private PlayStyle myStyle;
-    private PlayCondition myStatus;
-    private VoiceChat myVoice;
-}
