--- src/main/java/com/tools/seoultech/timoproject/matching/domain/user/entity/enumType/PlayCondition.java
@@ -1,7 +1,0 @@
-package com.tools.seoultech.timoproject.matching.domain.user.entity.enumType;
-
-public enum PlayCondition {
-    FIRST,
-    CONTINUE,
-    LAST
-}
