--- src/main/java/com/tools/seoultech/timoproject/matching/domain/user/entity/enumType/VoiceChat.java
@@ -1,6 +1,0 @@
-package com.tools.seoultech.timoproject.matching.domain.user.entity.enumType;
-
-public enum VoiceChat {
-    ENABLED,
-    DISABLED,
-}
