+++ src/main/java/com/tools/seoultech/timoproject/ranking/dto/RedisRankingInfo.java
@@ -2,8 +2,8 @@ package com.tools.seoultech.timoproject.ranking.dto;
 
 
 import com.fasterxml.jackson.annotation.JsonFormat;
-import com.tools.seoultech.timoproject.matching.domain.user.entity.enumType.Gender;
-import com.tools.seoultech.timoproject.matching.domain.user.entity.enumType.PlayPosition;
+import com.tools.seoultech.timoproject.matching.domain.board.entity.enumType.Gender;
+import com.tools.seoultech.timoproject.matching.domain.board.entity.enumType.PlayPosition;
 import com.tools.seoultech.timoproject.riot.dto.RiotRankingDto;
 import com.tools.seoultech.timoproject.memberAccount.domain.entity.Member;
 import jakarta.persistence.Id;
